requirements.txt
Django==2.2
django-environ==0.9.0
django-filter==21.1
djangorestframework==3.13.1
pytz==2022.6
sqlparse==0.4.3
